Don't do an HTTP request when retrieving URL
The `#public_url` which is the only method documented on every storage,
but it makes and HTTP request, and it seems to do that for every storage
concistently. All storages that I've seen also implement the `#url`
method which seems to be what we want. It does require an `expires`
argument, but people can just set it to 20 years if they want.